comments: archive empty suggestion branches on open; rename Suggestions/; changeset
- Archive lifecycle: sweep suggestion branches with no changes vs their fork point when
  the companion opens (not after each accept/reject, so the reject op's undo isn't raced).
- Rename components/SuggestionCard/ -> components/Suggestions/ (folder now holds SuggestionThread).
- Extend the document-suggestion-review changeset to cover the review layer (@dxos/types
  Change block, react-ui-thread tile, plugin-comments companion, plugin-markdown SuggestEdit).
- Track: demo-video test plan + usage script (document-revisions TASKS.md).
